Nuget package changes: Use a meta-package, rename to runtime, remove other DLLs and EXEs (#9)
This change makes a number of changes to the Nuget build scripts and configuration for MS-ICU.

It removes the all-in-one Nuget package approach that was used previously, and instead now uses a meta-package with dependencies. These dependent packages contain architecture specific DLLs, which can be built separately from the main meta-package.

The name of the package is changed to "Microsoft.ICU.icu4c.runtime" to reflect that it only contains the 'common' and 'i18n' code libraries from ICU4C, along with the data DLL. No other binaries are included in the package. (Ex: the ICU tools or the IO library).
The headers and libs/pdbs are also omitted as well, in order to make the "runtime" package as small as possible.