Statically link the VCRuntime, VCStartup, and STL, but dynamically link to the UCRT. (#10)
This change modifies the vcxproj files for the "common" (uc) and "i18n" (in) libraries in order to statically link the VCRuntime (libvcruntime.lib), VCStartup (libcmt.lib), and STL/MSVCP (libcpmt.lib), but dynamically link the UCRT (ucrt.lib).

This change also adds a new MSFT-Patch file to the icu-patches folder in order to retain this change for future versions of ICU. (I also reworded the readme.md file a bit for the patches folder).

The UCRT (ucrtbase.dll) is included as Windows OS component in Windows 10 and above. (See https://docs.microsoft.com/cpp/c-runtime-library/crt-library-features). On previous versions of Windows it is available via a Windows Update package. (See the KB article here: https://support.microsoft.com/kb/2999226).

However, the vcruntime*.dll and msvcp*.dll files are only available via the Visual C++ Redistributable (VC Redist), and the Redist must be installed manually.

This change avoids the need to install the VC Redist on Windows 10 (and previous Windows versions with the KB update) in order to use the icuuc*.dll and icuin*.dll libraries.

Changes included in this commit:
- Static link the VCRUNTIME, VCSTARTUP, and STL, but dynamic link to the UCRT.
- Add MSFT-Patch file for "Static linking the VCRUNTIME, VCSTARTUP, and STL, but Dynamic linking the UCRT for the "common" and "i18n" projects".
- Clarify the wording in the readme for the patches.