Convert the Nuget build pipeline to YAML, and check it in to the repo (#17)
This change converts the Nuget build pipeline configuration from the ADO "classic" configuration into a YML/YAML file, which can be checked in to the repository.

Previously the Nuget build pipeline configuration only existed on MSCodeHub, meaning that nobody else could make any changes to it. This meant that in order to (for example) add support for a new platform to the Nuget package, changes would need to be done manually by someone on our team to the ADO pipeline setup within MSCodeHub.
With this change, anyone can create a PR to propose changes to add support for a new platform, etc. to the Nuget package pipeline.

This change also fixes the casing on the Nuget name to be "Microsoft.ICU.ICU4C.Runtime".

Changes in this commit:
* Only add the alpha+buildnum to the Nuget package if prerelease is passed to the script.
* Fix the casing on the package name.
* Change the wording in the nuspec slightly.
* Add YAML build file for Nuget package.