Fix setting the MS-ICU version number in the uvernum.h file, and don't set it at build time. (#20)
When testing out the Nuget package, the value reported by the API `u_getVersion` doesn't
actually match the MS-ICU version number. This is because the macro `U_ICU_VERSION`
wasn't updated, only the macro `U_ICU_VERSION_BUILDLEVEL_NUM` was being
updated.

Initially, I was going to modify the `Set-ICUVersion.ps1` script to set this macro as well at
build time. However, when testing I noticed that the `make dist` command actually takes
the output from the git commit using the `git archive` command. This means that we
need to actually modify and commit the version number, and not set it on the fly during
the build.

Instead, we'll now have to update the two macros in the `uvernum.h` file directly, but in
order to make sure things don't get out of sync, I modified the `Set-ICUVersion.ps1`
script to check that they match.

Also in this change:
I added a "clean" step to the CI pipeline to clean everything, as I noticed that if you do a
force push, sometimes the contents of the build machines doesn't get cleaned properly.