Produce debugging symbols for Linux x64. (#30)
This change enables the creation of debugging symbols for the Linux x64 binaries that are part of the MS-ICU Nuget package.

In order to produce the debugging symbols for the Linux ELF binaries, we need to build the ICU library code with symbols support (ie: the `-g` compiler option).

We want to produce debugging symbols for "release" builds, but we don't want to use the `--enable-debug` flag/option with the `runConfigureICU` script, as that will turn on all kinds of debug code and asserts inside the ICU library code.

Instead we set the environment variables `CFLAGS` and `CXXFLAGS` (which are used by the compiler) to add the `-g` option. This lets us to do a full "release" build but still have symbols.

Debugging symbols are handled differently on Linux than Windows, and the debug symbols are actually *included* as part of the binary itself. (ie: This would be like if the `.dll` file also had all the info from the `.pdb` file included in it).

This means that once the library is built we need to strip off the debugging info from the `.so` binary into a separate file.
The convention for these "debugging symbols" files on Linux is to use the (aptly named) extension of "`.debug`".

The resulting symbol files are published as a build artifact named "`symbols-linux-x64`" on the CI pipeline.

To give a concrete example, the "linux-x64" artifact contains:
- `libicudata.so.67.1.0.5`
- `libicuuc.so.67.1.0.5`
- `libicui18n.so.67.1.0.5`

With this change, we'll now have a "symbols-linux-x64" artifact with the following files:
- `libicudata.so.67.1.0.5.debug`
- `libicuuc.so.67.1.0.5.debug`
- `libicui18n.so.67.1.0.5.debug`

This change also fixes the variable `CPUCORES` in the Docker build script, and removes the unneeded binary tarball.