Publish Windows symbols to the public Microsoft symbol server, after signing the Nuget. (#33)
This change modifies the Nuget build pipeline to publish the debugging symbols for the Window
pre-built binaries to the public Microsoft symbol server. This also moves the publishing step to be
after the code signing steps, in case they might fail. (If they fail we don't want to publish the symbols).

This means that the symbols for the pre-built Windows binaries in the Nuget package can now be
pulled from the public symbol server URLs:
- https://msdl.microsoft.com/download/symbols (external)
- http://symweb (internal)

Note: The authorization for this publishing step is based on the individual user that manually triggers
runs the pipeline run. This means that we don't need to use a special domain-connected service
account or setup Azure Key vault storage for a custom PAT token.

Since we can use the public symbol server now, I also removed the publishing to the MSCodeHub
artifact storage, since that seems redundant now.