[ICU 68.1 Update] Initial commit of unmodified ICU source code from maint-68 branch (#50)
This is part 1 of N for updating to ICU 68.1.

This is an initial commit of the unmodified ICU source code and data from the upstream `maint/maint-68` branch.

I'm doing this as it's own separate commit and PR in order to make the history cleaner, and to make reviewing subsequent changes easier. Otherwise the diffs would be so massive that GitHub can't even load them.

## Detailed Description

ICU 68.1 release notes: http://site.icu-project.org/download/68
CLDR 38: release notes: http://cldr.unicode.org/index/downloads/cldr-38

Note: The upstream `maint/maint-68` branch currently has additional fixes that were not part of the actual ICU 68.1 release.

However, I'm including them in this PR in order to avoid having to separately cherry-pick them later.
The ICU-TC is planning on having a "dot release" for 68 in order to fix issues that were identified after 68.1 was released. This 68.2 version should come out some time in Dec/Jan, and we'll need to update to this later. We can avoid a small bit of churn by pulling in the fixes from the upstream `maint/maint-68` branch now.

I've also been updating/revising the [Updating MS-ICU](https://osgwiki.com/wiki/Updating_MS-ICU) wiki page as I go in order to  make changes due to the GitHub undocking work and other changes.

This PR is going directly into the `master` branch, as we have separate branches for the `maint/maint-64` and `maint/maint-67` versions on GitHub. Once this is merged the `master` branch will no longer be "in sync" with 67.1, and will become a 68.1 "staging" branch. Similarly on the MSCodeHub side, I'll create a versioned branch for the MS-ICU 67.1 version, and then once the git submodule is updated the `master` branch will be the "staging" branch.

Next up will be the update to the CLDR-to-ICU data build tool. After that is the update/removal and application of the various MSFT patch files and version updates. However, they will need to all be regenerated by hand due to the migration to GitHub.

Note: The CI builds for this PR will _fail_ and that is _expected_. I will need to update the CI pipelines later on afterwards, but that needs to be done as a separate commit. For now I will bypass the checks.