[ICU 68.1 Update] Initial commit of unmodified CLDR-to-ICU build tool (#51)
This commit is part 2 of N for updating to ICU 68.1.

This is an initial commit of the unmodified CLDR-to-ICU data build tool source code from the upstream `maint/maint-68` branch.

I'm doing this as it's own separate commit and PR in order to make the history cleaner, and to make reviewing subsequent changes easier. Otherwise the diffs would be so massive that GitHub can't even load them.

Note: The CI builds for this PR will _fail_ and that is _expected_. I will need to update the CI pipelines later on afterwards, but that needs to be done as a separate commit. For now I will bypass the checks.