The "tools/cldr/.gitignore" shouldn't ignore already checked-in files.
It's generally considered a bad practice or an anti-pattern to have files
checked-into a git repo which are at the same time also excluded by
.gitignore file(s) in the repo as well.

These two files are used by the CLDR-to-ICU data build process. However,
I accidentally removed them in the previous commit/PR, as there is a
checked-in .gitignore file which excludes them. (This means when coping
from one repo/clone to another repo/clone the files won't get added.)

This commit adds them back, and also add an exclusion to the .gitignore
file so that they don't get ignored.

I filed an upstream ICU ticket for this here as well:
https://unicode-org.atlassian.net/browse/ICU-21427