[ICU 68.1 Update] Increase the build timeout for CI-Nuget builds (#58)
This is part 9 of N for updating to ICU 68.1.

This is a relatively minor change. I noticed that the ARM builds are now taking a bit longer, causing the CI build to occasionally timeout and require triggering a re-run. This change simply increases the timeout amount on the build CI config to allow more time for the ARM builds to finish.