[ICU 68.1 Update] Rebuild data using extra data from CLDR seed locales (#60)
This is part 11 of N for updating to ICU 68.1.

For the extra locales that we pull in from CLDR Seed (for parity with NLS, etc). we also want to pull in their data for annotations, casing, collation, rbnf, and transforms as well.

This change rebuilds the ICU data using the rebuilt CLDR-MS data from MSCodeHub.
See the related PR here in MSCodeHub:
https://mscodehub.visualstudio.com/OSS_github.com_Unicode-org/_git/ms-icu/pullrequest/13846

This change adds the collation data for the 5 seed locales (aa, nso, ssy, st, tn). which have collation data files in CLDR.

Note: I'm not adding any other locales in the `build-icu-data.xml` file, as they don't have any data in CLDR seed for collation. Doing so makes the data build tool generated empty/stub files for them, which I don't think we want.

### Notes
For these new collations, we'll need to add the seed locales to the `IcuCollationNameToGuidIncrementJumplist.h` header for the ICU sorting adapter.

*Regarding add or removing collation GUIDs for NLS:*
In the NLS sorting paradigm, the GUIDs are effectively _tied_ to a particular NLS version number. Applications that care about reindexing are supposed to check _both_ the Version number and the GUID. If either change, then they need to reindex things.
(Though it's slightly more nuanced, as the set of GUIDs is shared in the registry for NLS for all versions. However, for the ICU-based sorting this doesn't apply.)

So as long as the major NLS version number is bumped, then additional and/or removal of GUIDs is fine.