[ICU 68.2 Update] Pickup code changes for upstream ICU 68.2 release.
This is part 12 of N for updating to ICU 68.1/68.2.

This change picks up the upstream code changes for ICU 68.2 from the
"release-68-2" tag on GitHub.

The ICU 68.2 release is a "bug fix" only release (on top of ICU 68.1),
and does not include any API changes.

From the ICU download page:
 "ICU 68 was released on 2020-10-28. Maintenance release 68.2 includes some bug fixes."

Note that this commit does not pick up any of the data changes, as we'll
need to rebuild the data using CLDR-MS (which has our data overrides)
once that has been updated to CLDR 38.1 "dot" release.

The CLDR project also issued a bug fix release in coordination with ICU.
However, the CLDR bug fix release was versioned as 38.1, since CLDR releases
start with an implicit zero.
(i.e.: ICU 68.1 uses CLDR 38, and ICU 68.2 uses CLDR 38.1)

There will be another commit on top of this commit next to update the data
files under the icu4c/source/data/* folder with the CLDR 38.1 changes from
the CLDR-MS repo in MSCodeHub.

The code changes for this were taken from the delta between the initial
ingestion of ICU 68.1 at commit ceeb53ac786e3a9bcd2520af1d57f227b8baab90
and the "release-68-2" tag.

You can see the comparison on GitHub here:
https://github.com/unicode-org/icu/compare/ceeb53ac786e3a9bcd2520af1d57f227b8baab90...release-68-2