Cherry-pick: Fix Windows TZ offset when Automatic DST setting is OFF (#65)
This change cherry-picks the upstream fix for the issue of the Windows
Time Zone offset being wrong when Automatic DST setting is OFF in the Control Panel.

ICU-21465 Windows Time Zone offset is wrong when Automatic DST setting is OFF
https://unicode-org.atlassian.net/browse/ICU-21465

Cherry-picked from upstream commit: c6934ff7476200fc22abdf4300c94e2a3460c793

The upstream PR into master is here:
https://github.com/unicode-org/icu/pull/1465

The upstream PR into maint/maint-68 is here:
https://github.com/unicode-org/icu/pull/1543

Detailed Description:
This fixes a bug in the changes for ICU-13845, for handling when the "Automatic DST" setting is OFF in the Control Panel.

It turns out that the offset value for the calculated time zone when the "Automatic DST" setting is turn off is wrong. The sign for the time zone offset is flipped. (It's positive when it should be negative, and vice-versa).

I think this was perhaps due to confusion about the value for the UTC offset, how POSIX style offsets are handled, and what is reported by the Win32 API `GetDynamicTimeZoneInformation`. It seems that this issue was actually present in the original PR for ICU-13845 and it was missed in the reviews on the other PR for the ticket.