MSFT-PATCH: Don't use the extended ICU data for Windows OS ICU.
For the Windows OS build of ICU, we only have one data file
and we don't use the extended data at all. We make this function a no-op
in order to save a few cycles for perf, but more importantly so that
we don't try to load a versioned data file (ex: icudt68l.dat) after
already loading the non-versioned common data file.