MSFT-PATCH: Static link VCRUNTIME, VCSTARTUP, STL, but dynamic link the UCRT
This modifies the .vcxproj files for the "common" (uc) and "i18n" (in) libraries in
order to statically link the VCRuntime (libvcruntime.lib), VCStartup (libcmt.lib),
and STL/MSVCP (libcpmt.lib), but dynamically link to the UCRT (ucrt.lib).

The UCRT (C Runtime library: ucrtbase.dll) is included in Windows 10 and above as an OS
component. (See https://docs.microsoft.com/cpp/c-runtime-library/crt-library-features ).
On previous versions of Windows it is available via a Windows Update package.
(See the KB article here: https://support.microsoft.com/kb/2999226 ).

However, the vcruntime*.dll and msvcp*.dll files are only available via the
Visual C++ Redistributable (VC Redist), and the Redist must be installed manually.

This change avoids the need to install the VC Redist on Windows 10, and
previous Windows versions with the KB update for the UCRT, in order to use
the two icuuc*.dll and icuin*.dll libraries.