MSFT-PATCH: CLDR-to-ICU data tool changes for MS-ICU/Windows (Part 2)
Windows only searches for .exe executables by default, so we can't just
call "ant" and instead need to launch a cmd prompt to run it.

The "&quot;" changes are needed in order double-quote the arguments
that get passed to the JVM. Without this hack the arguments don't get
quoted at all, which causes build failures as some of the variables
which get replaced have embedded spaces in them.