Add uprefs library to ICU to obtain the default locale as a full  BCP47 tag (#112)
 Currently, for many processes and tasks, ICU gets the default locale and caches it. This means that when needed, ICU will get something like "en-US" and that will not change even if you were to change your language or region in your device.
Furthermore, ICU has currently no way of getting other globalization settings such as currency, calendar, hour cycle, first day of week, sorting method and measurement system.
We have decided to add a way to solve these two problems.
By adding the uprefs library (only to the Windows implementation of uprv_getDefaultLocaleID()), we are adding the Uprefs_getBCP47Tag() internal API, which obtains a full, canonical and valid BCP47Tag containing all of the settings.

This means we also change the way we get the default locale. We go from getting only the locale and region, to getting the full thing.